14.修改很多BUG，租用私教和衣柜减少余额等

+++ src/main/java/com/xidian/meiping/service/implement/CardServiceImpl.java
@@ -1,10 +1,12 @@
 package com.xidian.meiping.service.implement;
 
 import com.xidian.meiping.dao.CardMapper;
+import com.xidian.meiping.dao.MemberMapper;
 import com.xidian.meiping.entity.Card;
 import com.xidian.meiping.service.service.CardService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -13,6 +15,8 @@ public class CardServiceImpl implements CardService {
 
     @Autowired
     private CardMapper cardMapper;
+    @Autowired
+    private MemberMapper memberMapper;
 
     @Override
     public List<Card> findAll() {
@@ -34,7 +38,10 @@ public class CardServiceImpl implements CardService {
         return cardMapper.selectByPrimaryKey(Id);
     }
     @Override
+    @Transactional
     public int deleteById(Integer Id) {
+        if(memberMapper.selectByCardTypeId(Id).size()>0)
+            return 0;
         return cardMapper.deleteByPrimaryKey(Id);
     }
