+++ src/main/java/com/xidian/meiping/service/implement/WardrobeServiceImpl.java
@@ -1,11 +1,14 @@
 package com.xidian.meiping.service.implement;
 
+import com.xidian.meiping.dao.MemberMapper;
 import com.xidian.meiping.dao.RendWardrobeLogMapper;
 import com.xidian.meiping.dao.WardrobeMapper;
+import com.xidian.meiping.entity.Member;
 import com.xidian.meiping.entity.RendWardrobeLog;
 import com.xidian.meiping.entity.Wardrobe;
 import com.xidian.meiping.service.service.WardrobeService;
 import com.xidian.meiping.util.CommonUtil;
+import com.xidian.meiping.util.StatusCode;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,7 +25,8 @@ public class WardrobeServiceImpl implements WardrobeService {
     private WardrobeMapper wardrobeMapper;
     @Autowired
     private RendWardrobeLogMapper mapper;
-
+    @Autowired
+    private MemberMapper memberMapper;
     @Override
     public List<Wardrobe> findAll() {
         return wardrobeMapper.selectAll();
@@ -31,12 +35,13 @@ public class WardrobeServiceImpl implements WardrobeService {
     @Override
     public int update(Wardrobe wardrobe) {
         wardrobe.setStartTime(null);
+        wardrobe.insertOrUpdate = true;
         return wardrobeMapper.updateByPrimaryKeySelective(wardrobe);
     }
 
     @Override
     public int add(Wardrobe wardrobe) {
-        System.out.println(wardrobe);
+        wardrobe.insertOrUpdate = true;
         //是否存在
         Wardrobe temp = wardrobeMapper.selectByPrimaryKey(wardrobe.getId());
 //        System.out.println(temp.toString());
@@ -54,8 +59,15 @@ public class WardrobeServiceImpl implements WardrobeService {
     @Override
     @Transactional
     public int deleteById(Integer Id) {
-        mapper.deleteByWardrobeId(Id);
-        return wardrobeMapper.deleteByPrimaryKey(Id);
+        Wardrobe wardrobe = wardrobeMapper.selectByPrimaryKey(Id);
+        System.out.println(wardrobe);
+        if(wardrobe.getStatus()== 1) return 0;
+        try{
+            return  wardrobeMapper.deleteByPrimaryKey(Id);
+        }catch (Exception e){
+            e.printStackTrace();
+            return 0;
+        }
     }
 
     @Override
@@ -85,7 +97,20 @@ public class WardrobeServiceImpl implements WardrobeService {
     @Override
     @Transactional
     public int rendWardrobe(HttpServletRequest request) {
+        double totalAmount = Double.parseDouble(request.getParameter("totalAmount"));
         Wardrobe wardrobe = (Wardrobe) CommonUtil.newInstance(new Wardrobe(),request);
+        wardrobe.insertOrUpdate = true;
+        Member member = memberMapper.selectByPrimaryKey(wardrobe.getMemberId());
+        member.insertOrUpdate = true;
+        if(member.getBalance()<totalAmount){
+            return StatusCode.BANLANCE_ERROR;
+        }
+        //修改会员余额
+        member.setBalance(member.getBalance()-totalAmount);
+        memberMapper.updateByPrimaryKeySelective(member);
+        //修改衣柜
+        wardrobeMapper.updateByMemberId(Integer.parseInt(request.getParameter("memberId")));
+        //插入衣柜租用记录
         RendWardrobeLog log = new RendWardrobeLog();
         log.setOperateTime(new SimpleDateFormat("yyyy-MM-dd HH:mm:ss")
                 .format(new Date()));
@@ -95,7 +120,7 @@ public class WardrobeServiceImpl implements WardrobeService {
         log.setWardrobeId(wardrobe.getId());
         log.setContext(request.getParameter("context"));
         log.setOperaterId(Integer.parseInt(request.getParameter("operaterId")));
-        log.setTotalAmount(Double.parseDouble(request.getParameter("totalAmount")));
+        log.setTotalAmount(totalAmount);
         mapper.insert(log);
 //        int id = mapper.selectByMemberIdAndWardrobeIdAndEndTime(log.getMemberId(),log.getWardrobeId(),log.getEndTime()).getId();
         wardrobe.setRendWardrobeLogId(log.getId());
